pyproject.toml
[tool.poetry]
name = "analyticaloan"
version = "1.0.0"
description = "AI-Powered Credit Underwriting System"
authors = ["AnalyticaLoan Team"]
readme = "README.md"
license = "Proprietary"

[tool.poetry.dependencies]
python = "^3.11"

# Web Framework
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# Database
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"

# Redis & Caching
redis = "^5.0.1"
hiredis = "^2.2.3"

# Authentication & Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
cryptography = "^41.0.7"

# HTTP Client
httpx = "^0.25.2"
aiohttp = "^3.9.1"

# Google Cloud & AI
google-cloud-aiplatform = "^1.38.1"
google-cloud-vision = "^3.4.5"
google-cloud-storage = "^2.13.0"
google-generativeai = "^0.3.0"

# ML & Data Science
scikit-learn = "^1.3.2"
xgboost = "^2.0.2"
lightgbm = "^4.1.0"
pandas = "^2.1.4"
numpy = "^1.26.2"
shap = "^0.44.0"

# MLflow for Model Management
mlflow = "^2.9.2"

# Vector Database
weaviate-client = "^3.25.3"

# Message Queue
pika = "^1.3.2"

# Workflow Engine
temporalio = "^1.4.0"

# Monitoring & Logging
prometheus-client = "^0.19.0"
structlog = "^23.2.0"
python-json-logger = "^2.0.7"

# PDF Processing
pypdf2 = "^3.0.1"
reportlab = "^4.0.7"
pillow = "^10.1.0"

# Utilities
python-dotenv = "^1.0.0"
tenacity = "^8.2.3"
pytz = "^2023.3"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
pytest-mock = "^3.12.0"
faker = "^20.1.0"

# Linting & Formatting
black = "^23.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
isort = "^5.13.2"
pylint = "^3.0.3"

# Type stubs
types-redis = "^4.6.0"
types-passlib = "^1.7.7"

# Development tools
ipython = "^8.18.1"
ipdb = "^0.13.13"

[tool.poetry.group.test.dependencies]
httpx = "^0.25.2"
pytest-httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=app --cov-report=term-missing --cov-report=html"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["services", "libs"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
